agents/job_hunting_agent/pyproject.toml
[project]
name = "job-hunting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "dotenv>=0.9.9",
    "pyjwt>=2.10.1",
    "a2a-sdk>=0.2.16",
    "google-adk>=1.0.0",
    "toolbox-core==0.3.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"